Add tests, finally
This aims to be a starting point for testing of this project and focuses
on the setup and simpler test cases. It's intended to be enough that,
from this point on, all changes include tests for affected parts.

Quick & Nimble is included via Carthage because Carthage is the simplest
possible approach and Quick & Nimble seems to be the clear choice for
BDD style testing.